move oauth support into the smtp lib, make sig type and email url configurable
