+++ lib/email/smtp.js
@@ -4,6 +4,7 @@ const {Cc, Ci, Cu} = require("chrome");
 const {Socket} = require("./socket");
 const {setTimeout} = require("timer");
 const {base64Encode, base64Decode} = require("api-utils/utils/data")
+const {OAuth} = require("OAuth");
 
 function encode_utf8( s )
 {
@@ -92,12 +93,39 @@ SslSmtpClient.prototype.onStartTls = function(reply, cb, cberr) {
     }
 };
 
+SslSmtpClient.prototype.build_xauth_string = function(xoauth) {
+    let accessor = {
+      consumerSecret: xoauth.consumerSecret,
+      consumerKey: xoauth.consumerKey,
+      tokenSecret: xoauth.tokenSecret,
+      token: xoauth.token
+    };
+    let action = xoauth.serviceProvider.emailUrl.replace('%s', this.smtpArgs.email);
+    let message = {method: 'GET', parameters: {
+            'oauth_signature_method': xoauth.serviceProvider.signatureMethod,
+            'oauth_token': xoauth.token
+      }, action: action};
+    OAuth.completeRequest(message, accessor);
+
+    // much like OAuth.getAuthorizationHeader except that does the realm etc.
+    let data = [];
+    let list = OAuth.getParameterList(message.parameters);
+    list.sort();
+    for (var p = 0; p < list.length; ++p) {
+      let value = list[p][1];
+      if (value == null) value = "";
+      data.push(OAuth.percentEncode(list[p][0]) + '="' + OAuth.percentEncode(value) + '"');
+    }
+    return "GET " + action + " " + data.join(",");
+}
+
 SslSmtpClient.prototype.authenticate = function(args, cb, cberr) {
-    let type, auth;
+    let type, auth, xoauth;
     if ('xoauth' in args && args.xoauth) {
         // We expect xoauth to hold the already escaped string with all the
         // oauth magic necessary - we just base64 encode it and pass it on.
-        auth = base64Encode(args.xoauth);
+        xoauth = this.build_xauth_string(args.xoauth)
+        auth = base64Encode(xoauth);
         type = "XOAUTH"
     } else if ('plain' in args) {
         type = "PLAIN"
